Refactor e-commerce backend structure, enhance functionality, and remove obsolete components
- Updated README.md to provide a clearer overview of the implemented features and project structure.
- Removed the obsolete super_mall.sql file, streamlining the database setup process.
- Introduced asynchronous processing capabilities with the addition of AsyncConfig and enabled scheduling in the main application.
- Deleted the Invoice management module, including its controller, service, and related entities, to simplify the codebase.
- Enhanced notification management by adding new endpoints for deleting and batch deleting notifications, along with improved notification entity structure.
- Updated payment processing to include a new PaymentCallback DTO for handling payment responses more effectively.
- Refactored various services and controllers to improve code clarity and maintainability.

These changes significantly improve the organization, functionality, and maintainability of the SuperMall e-commerce backend.

+++ src/main/java/com/supermall/backend/domain/log/aspect/SystemLogAspect.java
@@ -27,12 +27,12 @@ public class SystemLogAspect {
     public Object logAround(ProceedingJoinPoint joinPoint) throws Throwable {
         // 获取当前登录用户
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        Long authId = null;
+        Integer authId = null;
         
         if (authentication != null && authentication.isAuthenticated() 
                 && !"anonymousUser".equals(authentication.getPrincipal())) {
             try {
-                authId = Long.parseLong(authentication.getName());
+                authId = Integer.parseInt(authentication.getName());
             } catch (NumberFormatException e) {
                 // 如果转换失败，保持authId为null
             }
