+++ src/main/java/com/supermall/backend/domain/log/service/SystemLogService.java
@@ -7,12 +7,12 @@ public interface SystemLogService {
     /**
      * 记录系统日志
      */
-    void log(Long authId, String module, String action, String detail, String ipAddress, String userAgent);
+    void log(Integer authId, String module, String action, String detail, String ipAddress, String userAgent);
     
     /**
      * 获取用户的操作日志
      */
-    Page<SystemLog> getUserLogs(Long authId, int page, int size);
+    Page<SystemLog> getUserLogs(Integer authId, int page, int size);
     
     /**
      * 获取指定模块的操作日志
