+++ src/main/java/com/supermall/backend/domain/product/service/ProductSpecService.java
@@ -5,10 +5,10 @@ import com.supermall.backend.domain.product.dto.ProductSpecResponse;
 import java.util.List;
 
 public interface ProductSpecService {
-    ProductSpecResponse addSpec(Long productId, ProductSpecRequest request);
-    void deleteSpec(Long productId, Long specId);
-    void updateSpec(Long productId, Long specId, ProductSpecRequest request);
-    List<ProductSpecResponse> getProductSpecs(Long productId);
-    void batchAddSpecs(Long productId, List<ProductSpecRequest> requests);
-    void batchDeleteSpecs(Long productId, List<Long> specIds);
+    ProductSpecResponse addSpec(Integer productId, ProductSpecRequest request);
+    void deleteSpec(Integer productId, Integer specId);
+    void updateSpec(Integer productId, Integer specId, ProductSpecRequest request);
+    List<ProductSpecResponse> getProductSpecs(Integer productId);
+    void batchAddSpecs(Integer productId, List<ProductSpecRequest> requests);
+    void batchDeleteSpecs(Integer productId, List<Integer> specIds);
 }
